refactor: [#1268] separate UDP handlers into diferent modules
Following HTTP structure.
